refactor(back): Move run session WS handlers into service
