+++ apps/backend/src/app/modules/session/run/run-session.service.ts
@@ -2,7 +2,8 @@ import {
   BadRequestException,
   Inject,
   Injectable,
-  InternalServerErrorException
+  InternalServerErrorException,
+  Logger
 } from '@nestjs/common';
 import * as Sentry from '@sentry/nestjs';
 import { LeaderboardRun, User } from '@momentum/db';
@@ -10,6 +11,7 @@ import {
   ActivityType,
   CompatibleStyles,
   GamemodeStyles,
+  KillswitchType,
   runPath,
   RunValidationError,
   RunValidationErrorType,
@@ -19,7 +21,17 @@ import {
 } from '@momentum/constants';
 import { FileStoreService } from '../../filestore/file-store.service';
 import { XpSystemsService } from '../../xp-systems/xp-systems.service';
-import { CompletedRunDto, DtoFactory, XpGainDto } from '../../../dto';
+import {
+  CompletedRunDto,
+  CreateRunSessionDto,
+  DtoFactory,
+  RunSessionErrorDto,
+  RunSessionIdDto,
+  RunSessionResponseDto,
+  UpdateRunSessionDto,
+  XpGainDto
+} from '../../../dto';
+import { KillswitchService } from '../../killswitch/killswitch.service';
 import { EXTENDED_PRISMA_SERVICE } from '../../database/db.constants';
 import {
   ExtendedPrismaService,
@@ -39,14 +51,206 @@ import { LeaderboardRunsDbService } from '../../runs/leaderboard-runs-db.service
 
 @Injectable()
 export class RunSessionService {
+  private readonly logger = new Logger(RunSessionService.name);
+
   constructor(
     @Inject(EXTENDED_PRISMA_SERVICE) private readonly db: ExtendedPrismaService,
     private readonly fileStoreService: RunFileStoreService,
     private readonly valkey: ValkeyService,
     private readonly xpSystems: XpSystemsService,
     private readonly mapsService: MapsService,
-    private readonly leaderboardRunsDbService: LeaderboardRunsDbService
+    private readonly leaderboardRunsDbService: LeaderboardRunsDbService,
+    private readonly killswitch: KillswitchService
   ) {}
+
+  //#region Run Session (game WebSocket)
+
+  /**
+   * Start a new session for the user on a leaderboard, clearing any existing
+   * session for the same track first.
+   */
+  async createSession(
+    userID: number,
+    data: CreateRunSessionDto
+  ): Promise<RunSessionResponseDto | RunSessionErrorDto> {
+    // When run submission is disabled, refuse to start new sessions
+    if (this.killswitch.checkKillswitch(KillswitchType.RUN_SUBMISSION)) {
+      this.logger.warn(
+        `createSession: blocked by RUN_SUBMISSION killswitch (userID=${userID})`
+      );
+      return { error: 'Run submission is currently disabled' };
+    }
+
+    const leaderboardData = {
+      mapID: data.mapID,
+      gamemode: data.gamemode,
+      trackType: data.trackType,
+      trackNum: data.trackNum
+    };
+
+    if (!(await this.db.leaderboard.exists({ where: leaderboardData }))) {
+      this.logger.warn(
+        `createSession: leaderboard not found (userID=${userID}, mapID=${data.mapID}, gamemode=${data.gamemode}, trackType=${data.trackType}, trackNum=${data.trackNum})`
+      );
+      return { error: 'Leaderboard does not exist' };
+    }
+
+    const sessionKey = idKey(userID);
+    const sessionIDs = await this.valkey.lrange(sessionKey, 0, -1);
+    for (const sessionID of sessionIDs) {
+      const session = await this.valkey.hgetall(dataKey(sessionID));
+      if (
+        session &&
+        Number(session.userID) === userID &&
+        Number(session.trackType) === data.trackType &&
+        Number(session.trackNum) === data.trackNum
+      ) {
+        await Promise.all([
+          this.valkey.lrem(sessionKey, 0, sessionID),
+          this.valkey.del(dataKey(sessionID))
+        ]);
+      }
+    }
+
+    const id = await this.valkey.incr(COUNTER_KEY);
+    const createdAt = Date.now();
+    const createdAtISO = new Date(createdAt).toISOString();
+
+    await Promise.all([
+      this.valkey.lpush(sessionKey, id),
+      this.valkey.hset(dataKey(id), { userID, createdAt, ...leaderboardData }),
+      this.valkey.lpush(
+        timestampKey(id),
+        serializeTimestamp(1, 1, 0, createdAt)
+      )
+    ]);
+
+    this.logger.log(
+      `createSession: created session (sessionID=${id}, userID=${userID}, mapID=${data.mapID})`
+    );
+    return {
+      id,
+      userID,
+      createdAt: createdAtISO,
+      ...leaderboardData,
+      timestamps: [
+        { majorNum: 1, minorNum: 1, time: 0, createdAt: createdAtISO }
+      ]
+    };
+  }
+
+  /**
+   * Append a timestamp/split to an existing session.
+   */
+  async updateSession(
+    userID: number,
+    data: UpdateRunSessionDto
+  ): Promise<null | RunSessionErrorDto> {
+    const storedUserID = await this.valkey.hget(
+      dataKey(data.sessionID),
+      'userID'
+    );
+
+    if (!storedUserID || Number(storedUserID) !== userID) {
+      this.logger.warn(
+        `updateSession: invalid session (sessionID=${data.sessionID}, userID=${userID})`
+      );
+      return { error: 'Invalid session' };
+    }
+
+    await this.valkey.lpush(
+      timestampKey(data.sessionID),
+      serializeTimestamp(data.majorNum, data.minorNum, data.time, Date.now())
+    );
+
+    this.logger.log(
+      `updateSession: timestamp added (sessionID=${data.sessionID}, userID=${userID}, majorNum=${data.majorNum}, minorNum=${data.minorNum}, time=${data.time})`
+    );
+    return null;
+  }
+
+  /**
+   * Discard a session and its data entirely.
+   */
+  async invalidateSession(
+    userID: number,
+    data: RunSessionIdDto
+  ): Promise<null | RunSessionErrorDto> {
+    const storedUserID = await this.valkey.hget(
+      dataKey(data.sessionID),
+      'userID'
+    );
+
+    if (!storedUserID || Number(storedUserID) !== userID) {
+      this.logger.warn(
+        `invalidateSession: invalid session (sessionID=${data.sessionID}, userID=${userID})`
+      );
+      return { error: 'Invalid session' };
+    }
+
+    await Promise.all([
+      this.valkey.lrem(idKey(userID), 0, data.sessionID),
+      this.valkey.del(dataKey(data.sessionID)),
+      this.valkey.del(timestampKey(data.sessionID))
+    ]);
+
+    this.logger.log(
+      `invalidateSession: session deleted (sessionID=${data.sessionID}, userID=${userID})`
+    );
+    return null;
+  }
+
+  /**
+   * Mark a session ended and hand ownership of its data to the replay upload
+   * (HTTP), which finalises and deletes it.
+   */
+  async endSession(
+    userID: number,
+    data: RunSessionIdDto
+  ): Promise<null | RunSessionErrorDto> {
+    const storedUserID = await this.valkey.hget(
+      dataKey(data.sessionID),
+      'userID'
+    );
+
+    // This event and the replay upload (HTTP POST /session/run/:id/end) race and
+    // can arrive in either order. If the session is already gone the upload won
+    // the race and consumed it - that's the expected terminal state, so ack
+    // without warning rather than treating it as an invalid session.
+    if (!storedUserID) {
+      return null;
+    }
+
+    if (Number(storedUserID) !== userID) {
+      this.logger.warn(
+        `endSession: invalid session (sessionID=${data.sessionID}, userID=${userID})`
+      );
+      return { error: 'Invalid session' };
+    }
+
+    // Don't delete the session data/timestamps here: the replay upload still
+    // needs them to validate and process the run, and may not have arrived yet.
+    // Instead mark the session ended, drop it from the user's active list, and
+    // set a TTL so the data is reaped if the upload never arrives (e.g. the
+    // client crashed after ending but before uploading). completeSession owns
+    // the final deletion once it has processed the run.
+    await Promise.all([
+      this.valkey.lrem(idKey(userID), 0, data.sessionID),
+      this.valkey.hset(dataKey(data.sessionID), 'ended', 1),
+      this.valkey.expire(dataKey(data.sessionID), ENDED_SESSION_TTL_SECONDS),
+      this.valkey.expire(
+        timestampKey(data.sessionID),
+        ENDED_SESSION_TTL_SECONDS
+      )
+    ]);
+
+    this.logger.log(
+      `endSession: session ended, awaiting replay upload (sessionID=${data.sessionID}, userID=${userID})`
+    );
+    return null;
+  }
+
+  //#endregion
 
   //#region Complete Session
 
@@ -466,6 +670,13 @@ export class RunSessionService {
 
 //#region Valkey Keys
 
+const COUNTER_KEY = 'runsess:counter';
+
+// How long an ended session's data/timestamps are kept alive waiting for the
+// replay upload (HTTP) to arrive and process it, before being reaped. Generous
+// enough to cover a slow upload of a large replay on a poor connection.
+const ENDED_SESSION_TTL_SECONDS = 5 * 60;
+
 function idKey(userID: number): string {
   return `runsess:id:${userID}`;
 }
@@ -476,6 +687,15 @@ function dataKey(sessionID: string | number): string {
 
 function timestampKey(sessionID: string | number): string {
   return `runsess:ts:${sessionID}`;
+}
+
+function serializeTimestamp(
+  majorNum: number,
+  minorNum: number,
+  time: number,
+  createdAt: number
+): string {
+  return `${majorNum},${minorNum},${time},${createdAt}`;
 }
 
 function deserializeTimestamp(str: string): RunSessionTimestamp {
